requirements.txt
playwright
beautifulsoup4
deepdiff
apscheduler
requests
python-dotenv
apscheduler
lxml
fastapi
